Rename string.h to cura_string.h to avoid headers naming conflicts
